v3 release
git-svn-id: http://www.bci2000.org/svn/trunk@3112 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/SignalProcessing/FIR/FIRFilter.cpp
@@ -2,8 +2,25 @@
 // $Id$
 // Description: A finite impulse response (FIR) filter for temporal filtering.
 //
-// (C) 2000-2010, BCI2000 Project
-// http://www.bci2000.org
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////////////////
 #include "PCHIncludes.h"
 #pragma hdrstop
@@ -57,7 +74,7 @@ FIRFilter::Preflight( const SignalProperties& Input,
       { // Reduce output to a single element per block.
         Output.SetElements( 1 );
         // Adapt element unit to obtain correct time axis.
-        float gain = Output.ElementUnit().Gain();
+        double gain = Output.ElementUnit().Gain();
         Output.ElementUnit().SetGain( gain * Input.Elements() );
       } break;
