+++ src/contrib/SignalProcessing/SW/FBArteCorrection.cpp
@@ -12,6 +12,25 @@
 //              Feb 24, 2004, jm: Moved the TFBArteCorrection class into
 //              separate files.
 //
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////////////////
 #include "PCHIncludes.h"
 #pragma hdrstop
@@ -68,7 +87,7 @@ TFBArteCorrection::Initialize( const SignalProperties&, const SignalProperties& 
 {
   mArteChList.clear();
   for( int i = 0; i < Parameter( "ArteChList" )->NumValues(); ++i )
-    mArteChList.push_back( Parameter( "ArteChList" )( i ) - 1 );
+    mArteChList.push_back( static_cast<int>( Parameter( "ArteChList" )( i ) ) - 1 );
 
   mArteFactorList.clear();
   for( int i = 0; i < Parameter( "ArteFactorList" )->NumValues(); ++i )
@@ -87,8 +106,8 @@ TFBArteCorrection::Process( const GenericSignal& inSignal, GenericSignal& outSig
     if( mArteChList[ i ] >= 0 )
       for( int j = 0; j < ioSignal.Elements(); ++j )
       {
-        float controlSignal = ioSignal( i, j ),
-              arteSignal = ioSignal( mArteChList[ i ], j ) * mArteFactorList[ i ];
+        double controlSignal = ioSignal( i, j ),
+               arteSignal = ioSignal( mArteChList[ i ], j ) * mArteFactorList[ i ];
 
         switch( mArteMode )
         {
