+++ src/contrib/SignalProcessing/SW/SWFilter.h
@@ -10,6 +10,25 @@
 //              Feb 8, 2004, jm: Adaptations to changes in BCI2000 framework,
 //              minor reformulations, reformatting.
 //
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////////////////
 #ifndef SWFilterH
 #define SWFilterH
@@ -51,10 +70,10 @@ class TSWFilter : public GenericFilter
     int                  mAvgBufferSize;
     GenericSignal        mAvgBlockBuffer;  // AvgBlockBuffer[short SWCh][int AvgBufferSize]
     // Tc-correction variables
-    float                mTcFactor;
+    double               mTcFactor;
     std::vector<double>  mTcAk; // TcAk[SWCh+ArteCh]
     // artefact variables
-    std::vector<float>   mThresholdAmp,    // ThresholdAmp[short SWCh ]
+    std::vector<double>  mThresholdAmp,    // ThresholdAmp[short SWCh ]
                          mMinValue,        // MinValue[short SWCh]
                          mMaxValue;        // MaxValue[short SWCh]
 };
