+++ src/contrib/SignalProcessing/SW/SetBaseline.cpp
@@ -11,6 +11,25 @@
 //           minor reformulations, reformatting.
 //           Feb 24, 2004, jm: Moved the TSetBaseline class into separate files.
 //
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////////////////
 #include "PCHIncludes.h"
 #pragma hdrstop
@@ -82,7 +101,7 @@ TSetBaseline::Process( const GenericSignal& inSignal, GenericSignal& outSignal )
   if( State( baselineStateName ) )
     for( int i = 0; i < ioSignal.Channels(); ++i )
     {
-      float sum = 0;
+      double sum = 0;
       for( int j = 0; j < ioSignal.Elements(); ++j )
         sum += ioSignal( i, j );
       mBLSignal( i, 0 ) = mBLSignal( i, 0 ) * mBLSamples[ i ] + sum;
