+++ src/contrib/SignalSource/RDAClient/RDAClientADC.cpp
@@ -6,8 +6,25 @@
 // Description: A source class that interfaces to the BrainAmp RDA socket
 //              interface.
 //
-// (C) 2000-2010, BCI2000 Project
-// http://www.bci2000.org
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////////////////
 #include "PCHIncludes.h"
 #pragma hdrstop
@@ -21,7 +38,7 @@
 
 using namespace std;
 
-const float eps = 1e-20; // Smaller values are considered zero.
+const float eps = 1e-20f; // Smaller values are considered zero.
 
 // Register the source class with the framework.
 RegisterFilter( RDAClientADC, 1 );
@@ -121,7 +138,7 @@ RDAClientADC::Preflight( const SignalProperties&,
              << endl;
     else
     {
-      float sourceSamplingRate = 1e6 / preflightQueue.info().samplingInterval;
+      double sourceSamplingRate = 1e6 / preflightQueue.info().samplingInterval;
       if( Parameter( "SamplingRate" ) != sourceSamplingRate )
         bcierr << "The SamplingRate parameter must match "
                << "the setting in the recording software "
@@ -130,8 +147,7 @@ RDAClientADC::Preflight( const SignalProperties&,
 
       // Check whether block sizes are sub-optimal.
       size_t sampleBlockSize = Parameter( "SampleBlockSize" ),
-             sourceBlockSize =
-        preflightQueue.info().blockDuration / preflightQueue.info().samplingInterval;
+             sourceBlockSize = static_cast<size_t>( preflightQueue.info().blockDuration / preflightQueue.info().samplingInterval );
       if( sampleBlockSize % sourceBlockSize != 0 && sourceBlockSize % sampleBlockSize != 0 )
         bciout << "Non-integral ratio between source block size (" << sourceBlockSize << ")"
                << " and system block size (" << sampleBlockSize << "). "
