+++ src/core/SignalSource/SignalGenerator/SignalGeneratorADC.h
@@ -3,8 +3,25 @@
 // Author: schalk@wadsworth.org, juergen.mellinger@uni-tuebingen.de
 // Description: An ADC class for testing purposes.
 //
-// (C) 2000-2010, BCI2000 Project
-// http://www.bci2000.org
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////////////////
 #ifndef SIGNAL_GENERATOR_ADC_H
 #define SIGNAL_GENERATOR_ADC_H
@@ -28,7 +45,7 @@ class SignalGeneratorADC : public GenericADC
 
  private:
   // Configuration
-  float  mSamplingRate,
+  double mSamplingRate,
          mSineFrequency,
          mSineAmplitude,
          mNoiseAmplitude,
@@ -39,7 +56,7 @@ class SignalGeneratorADC : public GenericADC
          mSineChannelZ;
   bool   mModulateAmplitude;
   // Internal State
-  float  mAmplitudeX,
+  double mAmplitudeX,
          mAmplitudeY,
          mAmplitudeZ,
          mSinePhase;
