+++ src/core/Tools/cmdline/bci_stream2mat.cpp
@@ -3,10 +3,28 @@
 // Author: juergen.mellinger@uni-tuebingen.de
 // Description: See the ToolInfo definition below.
 //
-// (C) 2000-2010, BCI2000 Project
-// http://www.bci2000.org
+// $BEGIN_BCI2000_LICENSE$
+// 
+// This file is part of BCI2000, a platform for real-time bio-signal research.
+// [ Copyright (C) 2000-2011: BCI2000 team and many external contributors ]
+// 
+// BCI2000 is free software: you can redistribute it and/or modify it under the
+// terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+// 
+// BCI2000 is distributed in the hope that it will be useful, but
+//                         WITHOUT ANY WARRANTY
+// - without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+// A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <http://www.gnu.org/licenses/>.
+// 
+// $END_BCI2000_LICENSE$
 ////////////////////////////////////////////////////////////////////
 #include <iostream>
+#include <cstdio>
 #include <set>
 
 #include "bci_tool.h"
@@ -249,11 +267,11 @@ StreamToMat::WriteData( const GenericSignal& s )
       WriteFloat32( 0 );
   else
     for( StringSet::const_iterator i = mStateNames.begin(); i != mStateNames.end(); ++i )
-      WriteFloat32( mpStatevector->StateValue( i->c_str() ) );
+      WriteFloat32( static_cast<float32>( mpStatevector->StateValue( i->c_str() ) ) );
 
   for( int i = 0; i < s.Channels(); ++i )
     for( int j = 0; j < s.Elements(); ++j )
-      WriteFloat32( s( i, j ) );
+      WriteFloat32( static_cast<float32>( s( i, j ) ) );
   ++mDataCols;
 }
